Enhanced checking of the requirements...
I tried rendering a gif with subtitles and ran into an error like this:

```
convert.im6: no decode delegate for this image format `/tmp/magick-dmhxdNv4' @ error/constitute.c/ReadImage/544.
convert.im6: no images defined `gif:-' @error/convert.c/ConvertImageCommand/3044.
```

This is because `convert` doesn't get any input from `ffmpeg` and this
was caused because `ffmpeg` was not compiled with the `libass` flag.

So I added some extra dependency checking in the `checkRequirements`
call.

This does seem to add roughly 0.02s to the app, based on my
extensive (read: running the app with `time`, like an animal).

The resulting gif in all its glory:

![WAT](https://pile.pjaspers.com/bob_wat.gif)